% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildCohort.R
\name{buildCohort}
\alias{buildCohort}
\title{Create the exposure and outcome cohorts}
\usage{
buildCohort(connectionDetails, cdmDatabaseSchema, resultsDatabaseSchema,
  treatment, comparator, outComeId)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in
the DatabaseConnector package.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{resultsDatabaseSchema}{Schema name where intermediate data can be stored. You will need to
have write priviliges in this schema. Note that for SQL Server, this
should include both the database and schema name, for example
'cdm_data.dbo'.}

\item{treatment}{The name of the of treatment cohort that you want to creat, for
example treatment = c("BigToBig.sql) will creat treatment cohort of
patients who start with Biguanide and remain on it.}

\item{comparator}{The name of the comparator cohort that you want to create, for
example comparator = c("SulfToSulf.sql) will create comparator cohort
with respect to treatment cohort above.}

\item{outComeId}{The outcome Id of interest.}
}
\description{
Create the exposure and outcome cohorts
}
\details{
This function will create treatment, comparator and outcomes cohort as included in this package.
}
