% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCohort.R
\name{getCohort}
\alias{getCohort}
\title{Get Cohort}
\usage{
getCohort(connectionDetails, cdmDatabaseSchema, resultsDatabaseSchema,
  sqlFileName, targetDatabaseSchema, targetCohortTable, idOne, idTwo, idThree,
  idFour, idFive, idSix)
}
\arguments{
\item{connectionDetails}{The details of database schema.}

\item{cdmDatabaseSchema}{The details of CDM schema.}

\item{resultsDatabaseSchema}{The details of results schema.}

\item{sqlFileName}{Name of SQL file that will used to construct the cohort.}

\item{targetDatabaseSchema}{Name of target database schema.}

\item{targetCohortTable}{Name of target cohort table.}

\item{idOne}{Numeric value = 1 represents one of the drug combination.
This is representative of Treatment Cohort.}

\item{idTwo}{Numeric value = 2 representing one of the drug combination.
This is representative of the Comparator Cohort.}

\item{idThree}{Numeric value = 3 representing one of the outcome measure.
In case of DiabetesTxPathway this represents HbA1c}

\item{idFour}{Numeric value = 4 representing one of the outcome measure.
In case of DiabetesTxPathway this represents Myocardial Infraction}
}
\details{
The function get cohort can be used to generate the required cohort for
a given drug combination. You can use this function to generate your own cohort
based on given SQL query or use one of the SQL query provided. For DiabetesTxPathway
it is highly recomended to use the SQLs provided in this package.
}
\author{
Rohit Vashisht
}
