% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCohortTxPath.R
\name{getCohortTxPath}
\alias{getCohortTxPath}
\title{Get Cohort Tx Path}
\usage{
getCohortTxPath(connectionDetails, cdmDatabaseSchema, resultsDatabaseSchema,
  sqlFileName, targetDatabaseSchema, targetCohortTable, idOne)
}
\arguments{
\item{connectionDetails}{The details of database schema.}

\item{cdmDatabaseSchema}{The details of CDM schema.}

\item{resultsDatabaseSchema}{The details of results schema.}

\item{sqlFileName}{Name of SQL file that will used to construct the cohort.}

\item{targetDatabaseSchema}{Name of target database schema.}

\item{targetCohortTable}{Name of target cohort table.}

\item{idOne}{Numeric value = 1 represents one of the drug combination. This is
representative of Treatment Cohort.}
}
\details{
The function get cohort can be used to generate the required cohort for a given drug combination.
You can use this function to generate your own cohort based on given SQL query or use one of the
SQL query provided. For DiabetesTxPathway it is highly recomended to use the SQLs provided in this
package.
}
\author{
Rohit Vashisht
}
