% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTxPath.R
\name{getTxPath}
\alias{getTxPath}
\title{getTxPath}
\usage{
getTxPath(connectionDetails = connectionDetails,
  cdmDatabaseSchema = cdmDatabaseSchema,
  resultsDatabaseSchema = resultsDatabaseSchema, cdmVersion = cdmVersion,
  results_path = results_path)
}
\arguments{
\item{connectionDetails}{The connection details of the database.}

\item{cdmDatabaseSchema}{The name of cdm database schema.}

\item{resultsDatabaseSchema}{The name of results database schema.}

\item{cdmVersion}{The name of cdm version, should be 5}
}
\details{
This function can be used to construct treatment pathways for T2D
}
\author{
Rohit Vashisht
}
