% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCompleteStudy.R
\name{runCompleteStudy}
\alias{runCompleteStudy}
\title{runStudy}
\usage{
runCompleteStudy(connectionDetails = connectionDetails,
  cdmDatabaseSchema = cdmDatabaseSchema,
  resultsDatabaseSchema = resultsDatabaseSchema, cdmVersion = cdmVersion,
  numThread = numThread, results_path = results_path)
}
\arguments{
\item{connectionDetails}{The connection details of the database.}

\item{cdmDatabaseSchema}{The cdm database schema}

\item{resultsDatabaseSchema}{The results datavase schema}

\item{cdmVersion}{The cdm version, should be 5 only}

\item{numThread}{Number of threads to be used.}
}
\details{
This function can be used to perform the DiabetesTxPathway analysis
end-to-end. Just supply the details and leave it running overnight. Please note
the function will perform analysis if there are atleast 100 patients for each
of the drug group considered in the study.
}
\author{
Rohit Vashisht
}
