% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runStudy.R
\name{runStudy}
\alias{runStudy}
\title{runStudy}
\usage{
runStudy(connectionDetails, cdmDatabaseSchema, resultsDatabaseSchema,
  cdmVersion, outComeId, outComeName)
}
\arguments{
\item{connectionDetails}{The connection details of the database.}

\item{cdm_database_schema}{The name of your cdm database schema.}

\item{results_database_schema}{The name of your results database schema.}

\item{target_database_schema}{The name of the target database schema.}

\item{target_cohort_table}{The name of target cohort table.}

\item{numThread}{Number of threads to be used.}

\item{idOne}{Treatment cohort ID}

\item{idTwo}{Comparator Cohort ID}

\item{idThree}{Outcome cohort ID}

\item{idFour}{Outcome cohort ID}

\item{idFive}{Outcome cohort ID}

\item{idSix}{Outcome cohort ID}
}
\details{
This function can be used to perform the DiabetesTxPathway analysis
end-to-end. Just supply the details and leave it running overnight. Please note
the function will perform analysis if there are atleast 100 patients for each
of the drug group considered in the study.
}
\author{
Rohit Vashisht
}
