% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugEfficacyAnalysis.R
\name{drugEfficacyAnalysis}
\alias{drugEfficacyAnalysis}
\title{drugEfficacyAnalysis}
\usage{
drugEfficacyAnalysis(connectionDetails, cdmDatabaseSchema,
  resultsDatabaseSchema, cid2Rm, outCome, cdmVersion, treatment, comparator,
  numThread)
}
\arguments{
\item{connectionDetails}{The connection details.}

\item{cdmDatabaseSchema}{Name of CDM database schema.}

\item{resultsDatabaseSchema}{Name of results database schema.}

\item{cid2Rm}{Concept Ids to remove before matching.}

\item{outCome}{Outcome Id of interest, for which the analysis need to be performed.}

\item{cdmVersion}{Version of cdm, should be 5 only.}

\item{treatment}{Name of treatment drug.}

\item{comparator}{Name of comparator drug.}

\item{numThread}{Number of threads.}
}
\details{
This function can be used to perform the drug efficacy analysis. Very briefly, this function a)
obtain treatment and comparator cohorts from target schema b) perform patient-level propensity
score matching c) perform cox-proportional hazard modeling of the given outcome. Please note this
function can only performe analysis if there are > 150 patients in both treatment and comparator
cohorts. Right now I've hard coded these values. May be in the future we'll see how to generalize
this.
}
\author{
Rohit Vashisht
}
